requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.6.0
    # via pydantic
anyio==3.7.1
    # via
    #   fastapi
    #   httpcore
    #   starlette
    #   watchfiles
bcrypt==4.0.1
    # via -r requirements.in
black==23.10.1
    # via -r requirements.in
build==1.0.3
    # via pip-tools
cachetools==5.3.2
    # via google-auth
certifi==2023.7.22
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.1
    # via requests
click==8.1.7
    # via
    #   black
    #   pip-tools
    #   uvicorn
colorama==0.4.6
    # via
    #   build
    #   click
    #   pytest
    #   uvicorn
dnspython==2.4.2
    # via
    #   -r requirements.in
    #   email-validator
    #   pymongo
email-validator==2.1.0.post1
    # via pydantic
fastapi==0.104.0
    # via -r requirements.in
flake8==6.1.0
    # via -r requirements.in
google-api-core[grpc]==2.12.0
    # via
    #   google-cloud-core
    #   google-cloud-secret-manager
    #   google-cloud-storage
google-auth==2.23.3
    # via
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-storage
google-cloud-core==2.3.3
    # via google-cloud-storage
google-cloud-secret-manager==2.16.4
    # via -r requirements.in
google-cloud-storage==2.12.0
    # via -r requirements.in
google-crc32c==1.5.0
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.6.0
    # via google-cloud-storage
googleapis-common-protos[grpc]==1.61.0
    # via
    #   google-api-core
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.12.6
    # via google-cloud-secret-manager
grpcio==1.59.2
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.59.2
    # via google-api-core
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==0.18.0
    # via httpx
httptools==0.6.1
    # via uvicorn
httpx==0.25.0
    # via -r requirements.in
idna==3.4
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
iniconfig==2.0.0
    # via pytest
mccabe==0.7.0
    # via flake8
motor==3.3.1
    # via -r requirements.in
mypy-extensions==1.0.0
    # via black
packaging==23.2
    # via
    #   black
    #   build
    #   pytest
pathspec==0.11.2
    # via black
pip-tools==7.3.0
    # via -r requirements.in
platformdirs==3.11.0
    # via black
pluggy==1.3.0
    # via pytest
proto-plus==1.22.3
    # via google-cloud-secret-manager
protobuf==4.24.4
    # via
    #   google-api-core
    #   google-cloud-secret-manager
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
pyasn1==0.5.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.3.0
    # via google-auth
pycodestyle==2.11.1
    # via flake8
pydantic[email]==2.4.2
    # via
    #   -r requirements.in
    #   fastapi
    #   pydantic-settings
pydantic-core==2.10.1
    # via pydantic
pydantic-settings[dotenv]==2.0.3
    # via -r requirements.in
pyflakes==3.1.0
    # via flake8
pyjwt==2.8.0
    # via -r requirements.in
pymongo==4.5.0
    # via motor
pyproject-hooks==1.0.0
    # via build
pytest==7.4.3
    # via -r requirements.in
python-dotenv==1.0.0
    # via
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.6
    # via -r requirements.in
pyyaml==6.0.1
    # via uvicorn
requests==2.31.0
    # via
    #   -r requirements.in
    #   google-api-core
    #   google-cloud-storage
rsa==4.9
    # via google-auth
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
starlette==0.27.0
    # via fastapi
typing-extensions==4.8.0
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
urllib3==2.0.7
    # via requests
uvicorn[standard]==0.23.2
    # via -r requirements.in
watchfiles==0.21.0
    # via uvicorn
websockets==12.0
    # via uvicorn
wheel==0.41.3
    # via pip-tools
mongomock==4.0.0
    # via -r requirements.in
mongomock_motor==0.0.21
    # via -r requirements.in
coverage==7.3.2
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
